credentials: Use `AsRef<str>` instead of `Borrow<str>`
`Credentials` and `Token` do not take advantage of `Borrow<T>` trait's
constraint that a type which implements `Borrow<T>` behaves identical to
`T`:

<https://doc.rust-lang.org/std/borrow/trait.Borrow.html>

Due to this constraint, the number of implementors of `Borrow<T>` is
fewer than that of `AsRef<T>`, limiting the ability of the users of
`Credentials` and `Token`.

One of few examples where a type implements `Borrow<T>` but not
`AsRef<T>` is the blanket implementation `impl<T> Borrow<T> for T`, but
this is not the case for `Credentials` and `Token` because `str` is an
unsized type and `Credentials` and `Token` cannot hold it directly.

This is a breaking change since `Borrow<T>` is not a subtrait of
`AsRef<T>`, but the impact should not be too massive because any type
which implements `Borrow<T>` _can_ implement `AsRef<T>` and such a type
typically implements `AsRef<T>` in fact (this is at least the case for
all `std` types).